(PC-12926) Split admin support_view code
